Add full 64-bit support to compressed arrays

+++ array/zfp/cache1.h
@@ -52,7 +52,7 @@ public:
   {
     for (typename zfp::Cache<CacheLine>::const_iterator p = cache.first(); p; p++) {
       if (p->tag.dirty()) {
-        uint block_index = p->tag.index() - 1;
+        size_t block_index = p->tag.index() - 1;
         store.encode(codec, block_index, p->line->data());
       }
       cache.flush(p->line);
@@ -68,30 +68,30 @@ public:
   }
 
   // inspector
-  Scalar get(uint i) const
+  Scalar get(size_t i) const
   {
     const CacheLine* p = line(i, false);
     return (*p)(i);
   }
 
   // mutator
-  void set(uint i, Scalar val)
+  void set(size_t i, Scalar val)
   {
     CacheLine* p = line(i, true);
     (*p)(i) = val;
   }
 
   // reference to cached element
-  Scalar& ref(uint i)
+  Scalar& ref(size_t i)
   {
     CacheLine* p = line(i, true);
     return (*p)(i);
   }
 
   // copy block from cache, if cached, or fetch from persistent storage without caching
-  void get_block(uint block_index, Scalar* p, ptrdiff_t sx) const
+  void get_block(size_t block_index, Scalar* p, ptrdiff_t sx) const
   {
-    const CacheLine* line = cache.lookup(block_index + 1, false);
+    const CacheLine* line = cache.lookup((uint)block_index + 1, false);
     if (line)
       line->get(p, sx, store.block_shape(block_index));
     else
@@ -99,9 +99,9 @@ public:
   }
 
   // copy block to cache, if cached, or store to persistent storage without caching
-  void put_block(uint block_index, const Scalar* p, ptrdiff_t sx)
+  void put_block(size_t block_index, const Scalar* p, ptrdiff_t sx)
   {
-    CacheLine* line = cache.lookup(block_index + 1, true);
+    CacheLine* line = cache.lookup((uint)block_index + 1, true);
     if (line)
       line->put(p, sx, store.block_shape(block_index));
     else
@@ -129,8 +129,8 @@ protected:
   class CacheLine {
   public:
     // accessors
-    Scalar operator()(uint i) const { return a[index(i)]; }
-    Scalar& operator()(uint i) { return a[index(i)]; }
+    Scalar operator()(size_t i) const { return a[index(i)]; }
+    Scalar& operator()(size_t i) { return a[index(i)]; }
 
     // pointer to decompressed block data
     const Scalar* data() const { return a; }
@@ -181,17 +181,17 @@ protected:
     }
 
   protected:
-    static uint index(uint i) { return (i & 3u); }
+    static size_t index(size_t i) { return (i & 3u); }
     Scalar a[4];
   };
 
   // return cache line for i; may require write-back and fetch
-  CacheLine* line(uint i, bool write) const
+  CacheLine* line(size_t i, bool write) const
   {
     CacheLine* p = 0;
-    uint block_index = store.block_index(i);
-    typename zfp::Cache<CacheLine>::Tag tag = cache.access(p, block_index + 1, write);
-    uint stored_block_index = tag.index() - 1;
+    size_t block_index = store.block_index(i);
+    typename zfp::Cache<CacheLine>::Tag tag = cache.access(p, (uint)block_index + 1, write);
+    size_t stored_block_index = tag.index() - 1;
     if (stored_block_index != block_index) {
       // write back occupied cache line if it is dirty
       if (tag.dirty())
@@ -214,7 +214,7 @@ protected:
   // number of cache lines corresponding to size (or suggested size if zero)
   static uint lines(size_t bytes, size_t blocks)
   {
-    uint n = bytes ? uint((bytes + sizeof(CacheLine) - 1) / sizeof(CacheLine)) : lines(blocks);
+    uint n = bytes ? static_cast<uint>((bytes + sizeof(CacheLine) - 1) / sizeof(CacheLine)) : lines(blocks);
     return std::max(n, 1u);
   }
