+++ array/zfp/pointer1.h
@@ -5,7 +5,7 @@ public:
   const_pointer() : const_handle(0, 0) {}
 
   // constructor
-  explicit const_pointer(container_type* container, uint i) : const_handle(container, i) {}
+  explicit const_pointer(container_type* container, size_t i) : const_handle(container, i) {}
 
   // dereference pointer
   const_reference operator*() const { return const_reference(container, i); }
@@ -51,7 +51,7 @@ public:
   pointer() : const_pointer(0, 0) {}
 
   // constructor
-  explicit pointer(container_type* container, uint i) : const_pointer(container, i) {}
+  explicit pointer(container_type* container, size_t i) : const_pointer(container, i) {}
 
   // dereference pointer
   reference operator*() const { return reference(container, i); }
