+++ array/zfp/reference2.h
@@ -4,7 +4,7 @@ public:
   typedef typename container_type::value_type value_type;
 
   // constructor
-  explicit const_reference(container_type* container, uint i, uint j) : const_handle(container, i, j) {}
+  explicit const_reference(container_type* container, size_t i, size_t j) : const_handle(container, i, j) {}
 
   // inspector
   operator value_type() const { return get(); }
@@ -23,7 +23,7 @@ protected:
 class reference : public const_reference {
 public:
   // constructor
-  explicit reference(container_type* container, uint i, uint j) : const_reference(container, i, j) {}
+  explicit reference(container_type* container, size_t i, size_t j) : const_reference(container, i, j) {}
 
   // assignment
   reference operator=(const reference& r) { set(r.get()); return *this; }
