+++ array/zfparray.h
@@ -53,7 +53,7 @@ protected:
   explicit array(uint dims, zfp_type type, const zfp::array::header& header) :
     type(type),
     dims(dims),
-    nx((uint)header.size_x()), ny((uint)header.size_y()), nz((uint)header.size_z())
+    nx(header.size_x()), ny(header.size_y()), nz(header.size_z())
   {
     if (header.scalar_type() != type)
       throw zfp::exception("zfp array scalar type does not match header");
@@ -85,9 +85,9 @@ protected:
     nz = a.nz;
   }
 
-  zfp_type type;   // scalar type
-  uint dims;       // array dimensionality (1, 2, or 3)
-  uint nx, ny, nz; // array dimensions
+  zfp_type type;     // scalar type
+  uint dims;         // array dimensionality (1, 2, or 3)
+  size_t nx, ny, nz; // array dimensions
 };
 
 }
