+++ tests/testviews.cpp
@@ -100,7 +100,7 @@ int main(int argc, char* argv[])
 
   // 2D slice of a
   printf("\n2D slice\n");
-  uint z = rand(0, nv.size_z());
+  uint z = rand(0, (uint)nv.size_z());
   zfp::array3<double>::nested_view2 slice2(nv[z]);
   for (uint y = 0; y < slice2.size_y(); y++)
     for (uint x = 0; x < slice2.size_x(); x++) {
@@ -119,7 +119,7 @@ int main(int argc, char* argv[])
 
   // 1D slice of a
   printf("\n1D slice\n");
-  uint y = rand(0, slice2.size_y());
+  uint y = rand(0, (uint)slice2.size_y());
   zfp::array3<double>::nested_view1 slice1 = slice2[y];
   for (uint x = 0; x < slice1.size_x(); x++) {
     printf("%u %u %u: %g %g\n", x, y, z, (double)a(x, y, z), (double)slice1[x]);
